+++ DawnLib/src/API/Items/.ItemRegistrationHandler.cs
@@ -140,7 +140,7 @@ static class ItemRegistrationHandler
                 };
                 level.spawnableScrap.Add(spawnableItemWithRarity);
             }
-            SpawnWeightContext ctx = new(level.GetDawnInfo(), RoundManager.Instance.dungeonGenerator?.Generator?.DungeonFlow?.GetDawnInfo(), TimeOfDayRefs.GetCurrentWeatherEffect(level)?.GetDawnInfo() ??  null);
+            SpawnWeightContext ctx = new(level.GetDawnInfo(), RoundManager.Instance.dungeonGenerator?.Generator?.DungeonFlow?.GetDawnInfo(), TimeOfDayRefs.GetCurrentWeatherEffect(level)?.GetDawnInfo() ?? null);
             spawnableItemWithRarity.rarity = scrapInfo.Weights.GetFor(ctx.Moon!, ctx) ?? 0;
         }
     }