+++ Thunderstore/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added to the Query implementation of DawnTerminalCommands.
   - You can now run a DawnEvent after ending a query.
   - You can now set a condition on the query for whether it should continue even after typing the continue keyword.
+- Interiors are now able to add extra scrap upon their generation.
 
 ## v0.7.10